aerotri-web/backend/requirements.txt
# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Image processing
Pillow==10.2.0

# GPU monitoring
pynvml==11.5.0

# WebSocket
websockets==12.0

# Utils
pydantic==2.5.3
python-dateutil==2.8.2
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
